requirements.txt
aiolimiter==1.0.0
anyio==3.6.2
APScheduler==3.10.1
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==3.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
python-dotenv==1.0.0
python-telegram-bot==20.1
pytz==2023.3
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzlocal==5.0.1
urllib3==2.0.4
